Description: optimize font memory IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I69CGJ Feature or Bugfix: feature Binary Source:No Signed-off-by: lancer <haoshuo@huawei.com>

+++ interfaces/kits/common/text.h
@@ -513,7 +513,7 @@ protected:
                                        uint8_t letterSpace, uint16_t& lineNum,
                                        uint16_t& letterIndex,
                                        SizeSpan* sizeSpans);
-    uint8_t GetSpanFontIdBySize(uint8_t size);
+    uint16_t GetSpanFontIdBySize(uint8_t size);
     void InitSizeSpans();
 #if defined(ENABLE_SPANNABLE_STRING) && ENABLE_SPANNABLE_STRING
     TextStyle* textStyles_;
